requirements.txt
aioredis==1.3.1
APScheduler==3.7.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.9.0
botocore==1.14.6
cachetools==5.0.0
certifi==2019.11.28
cffi==1.15.0
channels==2.2.0
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
config==0.5.1
configparser==5.1.0
constantly==15.1.0
cos-python-sdk-v5==1.9.15
crcmod==1.7
cryptography==35.0.0
cssselect==1.1.0
cssutils==2.3.0
daphne==2.5.0
dbus-python==1.2.16
ddddocr==1.4.2
decorator==4.4.2
dicttoxml==1.7.4
distro==1.4.0
Django==3.0.3
django-apscheduler==0.5.2
django-debug-toolbar==3.2
django-extensions==3.1.3
django-redis==5.2.0
django-simpleui==2021.4.3
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
flatbuffers==2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipython==7.13.0
ipython-genutils==0.2.0
itchat==1.3.10
jedi==0.15.2
jmespath==0.10.0
lxml==4.7.1
msgpack==1.0.3
numpy==1.18.1
onnxruntime==1.10.0
opencv-python==4.5.5.62
pandas==0.25.3
parso==0.5.2
pexpect==4.6.0
phone==0.4.3
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.11
premailer==3.10.0
prompt-toolkit==2.0.10
protobuf==3.19.4
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
PyExecJS==1.5.1
Pygments==2.3.1
PyGObject==3.36.0
PyHDFS==0.3.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pypng==0.0.21
PyQRCode==1.2.1
pyspark==3.1.1
python-dateutil==2.8.2
pytz==2021.3
redis==3.4.1
requests==2.27.1
rsa==4.8
service-identity==21.1.0
simplejson==3.17.6
six==1.14.0
snownlp==0.12.3
soupsieve==2.3.1
sqlparse==0.4.2
ssh-import-id==5.10
tencentcloud-sdk-python==3.0.583
thrift==0.15.0
traitlets==4.3.3
twilio==7.7.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
ufw==0.36
urllib3==1.25.8
uWSGI==2.0.20
wcwidth==0.1.8
yagmail==0.14.247
yarg==0.1.9
zope.interface==5.4.0
